dont update last comments while page is invisible
